Exactly compute C_2 - C_1 for interval.
Previously I was trying to compute C_2 directly, but this led to a proliferation of cases. It's easier to compute C_2 - C_1: we know the distribution of the leftmost point that hasn't been double covered ("leftmost single covered" or LSC point), viewed at the time we first reach n. At this point we can simply compute the expected time to reach the LSC point i, computed as E_i(n-i)^2.